add gpg

+++ sofa-jobs/sofa-jobs-impl/src/main/java/me/izhong/jobs/manage/impl/JobMngImpl.java
@@ -349,9 +349,15 @@ public class JobMngImpl implements IJobMngFacade {
         return null;
     }
 
+    @Transactional
     @Override
     public JobScript addJobScript(JobScript script) {
         ZJobScript dbBean = jobScriptService.insert(JobScriptUtil.toDbBean(script));
+
+        Long jobScriptId = dbBean.getJobScriptId();
+        jobInfoService.updateJobScriptId(jobScriptId,jobScriptId);
+
+
         return JobScriptUtil.toRpcBean(dbBean);
     }
 
@@ -398,6 +404,7 @@ public class JobMngImpl implements IJobMngFacade {
         return JobGroupUtil.toRpcBean(db);
     }
 
+    @Transactional
     @Override
     public JobGroup updateJobGroup(JobGroup group) {
         ZJobGroup db = jobGroupService.update(JobGroupUtil.toDbBean(group));
@@ -409,6 +416,11 @@ public class JobMngImpl implements IJobMngFacade {
     public long removeJobGroup(List<Long> groupIds) {
         long c = 0;
         for(Long l:groupIds){
+            List<ZJobInfo> lls = jobInfoService.selectByJobGroupId(l);
+            if(lls != null && lls.size() > 0) {
+                ZJobInfo zJobInfo = lls.get(0);
+                throw BusinessException.build("任务组[" + zJobInfo.getJobGroupId() + "]删除失败,有任务[" + zJobInfo.getJobDesc() + "]在使用");
+            }
             c += jobGroupService.deleteByPId(l);
         }
         return c;
